modes: Factored out support for modal models into a separate file. #48
